+++ apps/@examples/checklist-custom/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @examples/checklist-custom
 
+## 0.0.17
+
+### Patch Changes
+
+- Updated dependencies [8bd8d1167]
+  - @dopt/react-checklist@7.0.0
+  - @dopt/react@3.3.0
+  - @dopt/react-users@0.2.0
+  - @dopt/react-rich-text@5.2.1
+
 ## 0.0.16
 
 ### Patch Changes