+++ apps/@examples/checklist/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @examples/checklist
 
+## 0.1.38
+
+### Patch Changes
+
+- Updated dependencies [8bd8d1167]
+  - @dopt/react-checklist@7.0.0
+  - @dopt/react@3.3.0
+  - @dopt/react-users@0.2.0
+
 ## 0.1.37
 
 ### Patch Changes