+++ create-dopt/template-example-vite/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @examples/**PACKAGE_NAME**
 
+## 0.0.39
+
+### Patch Changes
+
+- Updated dependencies [8bd8d1167]
+  - @dopt/react@3.3.0
+  - @dopt/react-users@0.2.0
+
 ## 0.0.38
 
 ### Patch Changes